moved *ServiceState* from api/service.go to api/attach.go
